Add 'Waiting for moderation' state.
This state is entered when the user has not previously been
moderated or vetted by another user. The vetting process is peer
based.
It is possible to not be vetted and still be allowed to make
a revision publication. This is a secondary form of vetting,
which trusts the publishing application/party.