Allow the collation process to be rerun
This provides a development feature that will allow
ruleset authors to rerun the collation process more than
once on a given book.

The removal of the collated content happens at the view layer
so that the collation procedure can remain true to form.
That is, it was only intended to be run once. And in production
it would only be run once.