Change republish logic following archive update_latest trigger
update_latest trigger only updates latest_modules when stateid is 1
(current).  When a binder is updated, the changed pages caused binders
to republish.  There is no need to republish an earlier version of the
binder in latest_modules that is being updated.  This is achieved by not
comparing the version in latest_modules and the one being published.