Rewrite post-publication as a general purpose channel processing utility (#155)
* Rewriting the post-publication proc as a general purpose channel proc

This will now be able to process more than one channel nofitification.
It also uses the componentized events. This means the logic is now separate
from the component using the logic, which means we have a better separate
of scope and smaller unit of work.

* Handle the error on failure to bake

* Switch to pytest for running tests

This switch is mostly being made because travis-ci is acting goofy when
running the tests. Other environments are able to run them just fine, but
travis-ci decides to do it's own thing; and of course travis-ci is really
difficult to debug.

* Add a channel processing startup event and subscriber

The event subscriber will load the notifications queue for anything that
may have had notification while the process was not running.

* Recover when a payload is not present

* Commit processing state during post-publication processing

* Add a message for users when an unknown exception occurs

* Handle known and unknown exceptions from the epub export