Unskip the celery tests in test_subscribers.py
There is one celery worker per test that use the same event loop, but
the event loop is terminated by the celery worker in the first test.
Subsequent tests try to access the event loop that is "closed" and
failed.

See https://github.com/celery/celery/issues/4088

A way to get around this issue is to force a new event loop to be
created in every test.