Unskip tests (#170)
* Move pyramid and celery app fixtures to conftest

Move scoped_pyramid_app and celery_app from test_subscribers.py to
conftest.py so they are available for other tests.

* Use the default test celery_app in scoped_pyramid_app

Set up the default celery_app pytest fixture with celery_config and
celery_parameters.  Use it in scoped_pyramid_app instead of creating our
own.

* Include database set up code in scoped_pyramid_app

This includes setting up the celery tables by creating a new session
manager for the celery app each time scoped_pyramid_app is used.

* Change AsyncResult exception testing to pytest.raises

AsyncResult.state seems to be SUCCESS even when there is an exception in
the test.

* Unskip the celery tests in test_subscribers.py

There is one celery worker per test that use the same event loop, but
the event loop is terminated by the celery worker in the first test.
Subsequent tests try to access the event loop that is "closed" and
failed.

See https://github.com/celery/celery/issues/4088

A way to get around this issue is to force a new event loop to be
created in every test.