Set "GOB - [servername]" as the title of content status page
The title was empty for the content status page which means that it is
difficult to find the tab amongst other open tabs in the browser.

This will use something like "GOB - archive.local.cnx.org" as the title
of the page so users knows which server the content status page is for.

Close #189