Remove PENDING checkbox from content status page
PENDING is the default state of a celery job even if the job does not
exist.  The way we have resolved this problem is to set the job state to
QUEUED right after creating the job.  So none of our jobs should be in
the PENDING state.

Remove the PENDING checkbox from the content status page.