Tidy up content status page and add "QUEUED" to filter (#197)
* Change status filter checkboxes to use the same input name

The checkboxes have input names like pending_filter, started_filter etc.
This changes the input name for all the checkboxes to status_filter, so
we can just get a list from `request.params.getall('status_filter')`.

* Move content status page inline css to main.css

It would be better to separate the styling from the content in html
template and put it into main.css so styles can potentially be shared.

This also changes some icons and colors for the different states.

* Set "GOB - [servername]" as the title of content status page

The title was empty for the content status page which means that it is
difficult to find the tab amongst other open tabs in the browser.

This will use something like "GOB - archive.local.cnx.org" as the title
of the page so users knows which server the content status page is for.

Close #189

* Add "QUEUED" to status filter on content status page

There is currently no way of filtering queued jobs, which is the state
before a job is started.

Add a "QUEUED" checkbox to allow user to filter queued jobs.

Close #193

* Remove PENDING checkbox from content status page

PENDING is the default state of a celery job even if the job does not
exist.  The way we have resolved this problem is to set the job state to
QUEUED right after creating the job.  So none of our jobs should be in
the PENDING state.

Remove the PENDING checkbox from the content status page.