Remove AsyncResult calls from content_status page (#212)
loading the AnsyncResults was causing the content-status page to take
almost a minute! to load for 100 results. By using the celery_task_meta
table directly, the entire page renders instantly. I suspect it's an
issue with the layers of abstraction: Celery is using sqlalchemy,
probably as a ORM, and AsnyResult returns a "heavy" object with more
info than we need. I did see lots and lots of round trips to the DB,
doing things like testing the session transaction level - so not only
round trips, but fresh connections! That may be more celery's fault than
sqlalchemy _per se_, but it's set my expectations for this sort of
thing. I timed it: 0.5 sec vs 50 sec. 100 fold - the page render time is
almost completely the AsnyResults lookups. This is a hack of a
workaround - proper fix will store more historic state data