Add requirement pinnings
This gives us the ability to pin to specific versions. See cnx-press as an
example of why this is good. The most direct benefit of this is when used
in relation to pyup.io. Our tests are run when a new dependency is released,
this is important, because we are then able to see breakage prior to deployment.