Correct the container's volume mount
Only mount the code portion rather than the whole directory and project.
This is done so that the built files can still be found by pip,
which has installed them in the CWD, because we've installed the package
as "editable". These built files will be missing if we mount the entire
project directory.