Add the wait-for commandline utility
This utility will wait for the connected services to become available
before executing the intended command.

This is specifically useful in the travis scenario where there is
a lack of user input delay. This will allow the database and queue
components to start before running the tests that utilize these resources.