Remove redundant mention of the cnx-epub dependency
Only one of the mentioned occurances works, which is problematic since they
aren't quite the same. The first mention was `cnx-epub` and the other is
`cnx-epub[collation]`. We need the collation requirements for this package.
Removing the first non-extra case resolves a dependency acquisition issue
where the collation dependencies where not being resolved.