Filter out invalid requirements that start with # or -
We ignore `#` and `-` requirement.txt lines because they aren't setuptool
compatible. The `#` is used for a comments. The `-` prefixes an install
option. For example, `-e` or `-r` are valid to pip but not to setuptools.

The `-e` requirement line should only be used for development, so ensuring
removing these entries from the dependency list shouldn't be a problem
if used correctly.