Remove upload pypi step in Jenkinsfile
We continue to have problems with jenkins-dev because of the filesystem being
full.  This means either poking devops or uploading to pypi manually.  To avoid
this problem, we are going to use this concourse pipeline to upload to pypi
instead:

https://github.com/openstax/concourse-pipelines/tree/master/pypi-dist-upload

Concourse is better at cleaning up so hopefully we won't see the same problem.